+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file documents any relevant changes done to ViUR html5 since version 2.
 This is the current development version.
 
 - Feature: Allow to bind multiple [name] attributes to one widget in HTML parser
+- Feature: `html5.utils.textToHtml()` extended to `clear`-parameter
 
 ## [2.5.0] Vesuv